Gradient / Tracking results
For edge identification, I changed the kernel from that in 'Circle detection / Video capture' commit to four kernel passes of derivative filters (a y-derivative and its negative, an x-derivative and its negative) and then the norm (square root of the sum of squares) of these arrays.
I'm still not sure which generates better results.